+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/_components/HomeContent.tsx
@@ -7,6 +7,7 @@ import WishListHeart from "./WishListHeart";
 import { EggClub } from "@/types/cardlist.types";
 import { MemberInfo } from "@/types/user.types";
 import { EggDayWithEggDayMember } from "@/types/eggday.types";
+import { formatterAge, formatterGender, formatterPeopleLimit } from "../../../_utils/formatter";
 
 interface HomeContentProps {
   clubInfo: EggClub;
@@ -25,43 +26,6 @@ const HomeContent = ({
   notificationData,
   stringCategory
 }: HomeContentProps) => {
-  // 성별 커스텀
-  const gender = (gender: null | string) => {
-    if (gender === null) {
-      return "누구나";
-    } else if (gender === "남성") {
-      return "남성만";
-    } else {
-      return "여성만";
-    }
-  };
-
-  // 인원 제한 커스텀
-  const limited = (limit: number) => {
-    if (limit === 100) {
-      return "최대 100명";
-    } else {
-      return `최대 ${limit}명`;
-    }
-  };
-
-  // 나이 커스텀
-  const age = (age: number) => {
-    if (age === 100) {
-      return "제한 없음";
-    } else if (age === 50) {
-      return "50대";
-    } else if (age === 49) {
-      return "40대";
-    } else if (age === 39) {
-      return "30대";
-    } else if (age === 29) {
-      return "20대";
-    } else if (age === 19) {
-      return "10대";
-    }
-  };
-
   return (
     <div className="flex flex-col items-center justify-center">
       <div className="flex overflow-hidden w-[390px] h-[332px] relative bg-gray-100 mb-6">
@@ -76,8 +40,6 @@ const HomeContent = ({
 
       <div className="w-full flex-col justify-start items-start gap-8 px-4 inline-flex">
         <div className="self-stretch flex-col justify-start items-start gap-5 flex">
-          {/* 여기  */}
-
           <div className="self-stretch flex-col justify-start items-start gap-1 flex">
             <div className="px-2 py-0.5 bg-neutral-800 rounded-[124px] justify-center items-center inline-flex">
               <Text variant="body-10" className="text-white">
@@ -111,21 +73,19 @@ const HomeContent = ({
           <div className="self-stretch h-40 flex-col justify-start items-start gap-2 flex">
             <div className="self-stretch justify-start items-center gap-4 inline-flex">
               <Text variant="subtitle-14">나이</Text>
-              <Text variant="body-14">{age(clubInfo.egg_club_age)}</Text>
+              <Text variant="body-14">{formatterAge(clubInfo.egg_club_age)}</Text>
             </div>
             <div className="self-stretch justify-start items-center gap-4 inline-flex">
               <Text variant="subtitle-14">성별</Text>
-              <Text variant="body-14"> {gender(clubInfo.egg_club_gender)}</Text>
+              <Text variant="body-14"> {formatterGender(clubInfo.egg_club_gender)}</Text>
             </div>
             <div className="self-stretch justify-start items-center gap-4 inline-flex">
               <Text variant="subtitle-14">인원</Text>
-              <Text variant="body-14">{limited(clubInfo.egg_club_people_limited)}</Text>
+              <Text variant="body-14">{formatterPeopleLimit(clubInfo.egg_club_people_limited)}</Text>
             </div>
           </div>
         </div>
         <div className="self-stretch h-[0px] border border-solid border-gray-50"></div>
-
-        {/* 여기 */}
 
         <div className="self-stretch flex-col justify-start items-start gap-4 flex">
           <div className="self-stretch h-6 flex-col justify-start items-start gap-2 flex">
