+++ src/app/(pages)/(club)/club/regular-club-sub/[id]/page.tsx
@@ -16,7 +16,6 @@ const OneTimeClubSubPage = async ({ params }: { params: { id: string } }) => {
 
   // 클럽 정보만 추출
   const clubInfo = memberData[0]?.egg_club;
-  // console.log("클럽인포:", clubInfo);
 
   // 일치하는 카테고리 찾기
   const matchCategory = subCategories.find((category) => category.sub_category_id === clubInfo.sub_category_id);
@@ -40,7 +39,6 @@ const OneTimeClubSubPage = async ({ params }: { params: { id: string } }) => {
       </div>
 
       <TabLayout>
-        {/* props를 통해 데이터 전달 */}
         <HomeContent
           clubInfo={clubInfo}
           hostInfo={hostInfo}
